// src/shared/api/generated/models/crearCategoriaDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Cancha Nica API
 * Contrato HTTP de Cancha Nica. Errores: application/problem+json (RFC 7807) con `code` del catálogo (schema ProblemDetailsDto).
 * OpenAPI spec version: 0.0.1
 */

export interface CrearCategoriaDto {
  /**
     * @minimum 5
     * @maximum 80
     * @nullable
     */
  edadMaxima?: number | null;
  /**
     * @minimum 5
     * @maximum 80
     * @nullable
     */
  edadMinima?: number | null;
  /** @maxLength 60 */
  nombre: string;
}